requirements.txt
Django
tensorflow  # or any other ML framework you used for the CIFAR-10 model
pillow